feat: implementa classe TransacaoDTO e refatora DespesaDTO e ReceitaDTO para herança
