+++ backend/src/main/java/br/com/gestorfinanceiro/dto/despesa/DespesaDTO.java
@@ -1,59 +1,22 @@
 package br.com.gestorfinanceiro.dto.despesa;
 
-import jakarta.validation.constraints.DecimalMin;
+import br.com.gestorfinanceiro.dto.common.TransacaoDTO;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
+public class DespesaDTO extends TransacaoDTO {
+    @NotBlank(message = "O destino do pagamento é obrigatória.")
+    private String destinoPagamento;
 
-public class DespesaDTO {
-
-    @NotNull(message = "A data é obrigatória.")
-    private LocalDate data;
-
+    // Sobrescreva para aplicar validações específicas
+    @Override
     @NotBlank(message = "A categoria é obrigatória.")
     @Pattern(
             regexp = "ALIMENTACAO|MORADIA|TRANSPORTE|LAZER",
             message = "Categoria inválida. Valores permitidos: ALIMENTACAO, MORADIA, TRANSPORTE, LAZER."
     )
-    private String categoria;
-
-    @NotNull(message = "O valor é obrigatório.")
-    @DecimalMin(value = "0.01", message = "O valor deve ser maior que zero.")
-    private BigDecimal valor;
-
-    @NotBlank(message = "O destino do pagamento é obrigatória.")
-    private String destinoPagamento;
-
-    @NotBlank(message = "As observações são obrigatórias.")
-    private String observacoes;
-
-    private String uuid;
-
-    public LocalDate getData() {
-        return data;
-    }
-
-    public void setData(LocalDate data) {
-        this.data = data;
-    }
-
     public String getCategoria() {
-        return categoria;
-    }
-
-    public void setCategoria(String categoria) {
-        this.categoria = categoria;
-    }
-
-    public BigDecimal getValor() {
-        return valor;
-    }
-
-    public void setValor(BigDecimal valor) {
-        this.valor = valor;
+        return super.getCategoria();
     }
 
     public String getDestinoPagamento() {
@@ -63,16 +26,4 @@ public class DespesaDTO {
     public void setDestinoPagamento(String destinoPagamento) {
         this.destinoPagamento = destinoPagamento;
     }
-
-    public String getObservacoes() {
-        return observacoes;
-    }
-
-    public void setObservacoes(String observacoes) {
-        this.observacoes = observacoes;
-    }
-
-    public String getUuid() {return uuid;}
-
-    public void setUuid(String uuid) {this.uuid = uuid;}
 }
